# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

/**
 * The type of document format to apply
 */
export type FormatRequestDocumentType =
  (typeof FormatRequestDocumentType)[keyof typeof FormatRequestDocumentType];

export const FormatRequestDocumentType = {
  academic: "academic",
  business: "business",
  apa: "apa",
  mla: "mla",
  chicago: "chicago",
  general: "general",
} as const;

export interface FormatRequest {
  /** The raw document text to format */
  text: string;
  /** The type of document format to apply */
  documentType: FormatRequestDocumentType;
  /** Optional document title */
  title?: string;
  /** Optional author name */
  author?: string;
  /** Optional date string */
  date?: string;
}

export interface FormatTextResponse {
  formattedText: string;
  documentType: string;
}

export interface ExtractTextResponse {
  /** The extracted plain text content */
  text: string;
  /** Original filename */
  filename: string;
  /** Detected file type (docx, pdf, txt) */
  fileType: string;
}

export interface ErrorResponse {
  error: string;
  message: string;
}

export type ExtractTextBody = {
  file: Blob;
};