# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: format
    description: Document formatting operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /format/docx:
    post:
      operationId: formatDocx
      tags: [format]
      summary: Format document as DOCX
      description: Formats raw text into a properly structured Word document
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FormatRequest"
      responses:
        "200":
          description: DOCX file binary
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /format/pdf:
    post:
      operationId: formatPdf
      tags: [format]
      summary: Format document as PDF
      description: Formats raw text into a properly structured PDF document
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FormatRequest"
      responses:
        "200":
          description: PDF file binary
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /format/txt:
    post:
      operationId: formatTxt
      tags: [format]
      summary: Format document as plain text
      description: Formats raw text into a properly structured plain text document
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FormatRequest"
      responses:
        "200":
          description: Formatted plain text
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FormatTextResponse"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /format/extract:
    post:
      operationId: extractText
      tags: [format]
      summary: Extract text from uploaded document
      description: Extracts plain text from an uploaded .docx, .pdf, or .txt file
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
              required:
                - file
      responses:
        "200":
          description: Extracted text content
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExtractTextResponse"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    FormatRequest:
      type: object
      properties:
        text:
          type: string
          description: The raw document text to format
        documentType:
          type: string
          enum: [academic, business, apa, mla, chicago, general]
          description: The type of document format to apply
        title:
          type: string
          description: Optional document title
        author:
          type: string
          description: Optional author name
        date:
          type: string
          description: Optional date string
        fontFamily:
          type: string
          enum: [Times New Roman, Arial, Georgia, Courier New]
          description: Font family for the document
        fontSize:
          type: integer
          enum: [10, 11, 12, 14, 16]
          description: Font size in points
        lineSpacing:
          type: string
          enum: [single, "1.5", double]
          description: Line spacing style
      required:
        - text
        - documentType
    FormatTextResponse:
      type: object
      properties:
        formattedText:
          type: string
        documentType:
          type: string
      required:
        - formattedText
        - documentType
    ExtractTextResponse:
      type: object
      properties:
        text:
          type: string
          description: The extracted plain text content
        filename:
          type: string
          description: Original filename
        fileType:
          type: string
          description: Detected file type (docx, pdf, txt)
      required:
        - text
        - filename
        - fileType
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required:
        - error
        - message