# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * Formats raw text into a properly structured Word document
 * @summary Format document as DOCX
 */
const formatTypingFields = {
  fontFamily: zod
    .enum(["Times New Roman", "Arial", "Georgia", "Courier New"])
    .optional()
    .describe("Font family for the document"),
  fontSize: zod
    .union([zod.literal(10), zod.literal(11), zod.literal(12), zod.literal(14), zod.literal(16)])
    .optional()
    .describe("Font size in points"),
  lineSpacing: zod
    .enum(["single", "1.5", "double"])
    .optional()
    .describe("Line spacing style"),
};

export const FormatDocxBody = zod.object({
  text: zod.string().describe("The raw document text to format"),
  documentType: zod
    .enum(["academic", "business", "apa", "mla", "chicago", "general", "edd_dissertation"])
    .describe("The type of document format to apply"),
  title: zod.string().optional().describe("Optional document title"),
  author: zod.string().optional().describe("Optional author name"),
  date: zod.string().optional().describe("Optional date string"),
  ...formatTypingFields,
});

/**
 * Formats raw text into a properly structured PDF document
 * @summary Format document as PDF
 */
export const FormatPdfBody = zod.object({
  text: zod.string().describe("The raw document text to format"),
  documentType: zod
    .enum(["academic", "business", "apa", "mla", "chicago", "general", "edd_dissertation"])
    .describe("The type of document format to apply"),
  title: zod.string().optional().describe("Optional document title"),
  author: zod.string().optional().describe("Optional author name"),
  date: zod.string().optional().describe("Optional date string"),
  ...formatTypingFields,
});

/**
 * Formats raw text into a properly structured plain text document
 * @summary Format document as plain text
 */
export const FormatTxtBody = zod.object({
  text: zod.string().describe("The raw document text to format"),
  documentType: zod
    .enum(["academic", "business", "apa", "mla", "chicago", "general", "edd_dissertation"])
    .describe("The type of document format to apply"),
  title: zod.string().optional().describe("Optional document title"),
  author: zod.string().optional().describe("Optional author name"),
  date: zod.string().optional().describe("Optional date string"),
  ...formatTypingFields,
});

export const FormatTxtResponse = zod.object({
  formattedText: zod.string(),
  documentType: zod.string(),
});

/**
 * Extracts plain text from an uploaded .docx, .pdf, or .txt file
 * @summary Extract text from uploaded document
 */
export const ExtractTextBody = zod.object({
  file: zod.instanceof(File),
});

export const ExtractTextResponse = zod.object({
  text: zod.string().describe("The extracted plain text content"),
  filename: zod.string().describe("Original filename"),
  fileType: zod.string().describe("Detected file type (docx, pdf, txt)"),
});