# lib/api-zod/src/generated/types/extractTextResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ExtractTextResponse {
  /** The extracted plain text content */
  text: string;
  /** Original filename */
  filename: string;
  /** Detected file type (docx, pdf, txt) */
  fileType: string;
}