# lib/api-zod/src/generated/types/formatRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { FormatRequestDocumentType } from "./formatRequestDocumentType";

export interface FormatRequest {
  /** The raw document text to format */
  text: string;
  /** The type of document format to apply */
  documentType: FormatRequestDocumentType;
  /** Optional document title */
  title?: string;
  /** Optional author name */
  author?: string;
  /** Optional date string */
  date?: string;
  /** Font family for the document */
  fontFamily?: "Times New Roman" | "Arial" | "Georgia" | "Courier New";
  /** Font size in points */
  fontSize?: 10 | 11 | 12 | 14 | 16;
  /** Line spacing style */
  lineSpacing?: "single" | "1.5" | "double";
}